Fix double Close() panic
When using the App controller or the Client controller, if the Close()
function is called twice or more, then the second call will trigger a
panic as the error channel is already closed.

This commit puts the channel at nil and checks if it wasn't already closed
before closing it.